feat : shadcn monorepo 셋팅
