+++ apps/client/package.json
@@ -13,6 +13,9 @@
   },
   "dependencies": {
     "@wepl/ui": "*",
+    "autoprefixer": "^10.4.19",
+    "postcss": "^8.4.38",
+    "tailwindcss": "^3.4.3",
     "undici": "^5.0.0"
   },
   "devDependencies": {
