+++ apps/client/src/app/appLayout.tsx
@@ -1,11 +1,7 @@
-import './lib/globals.css';
+import '@ui/style/globals.css';
 import { QueryProvider } from './queryProvider';
 
-export function AppLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}): JSX.Element {
+export function AppLayout({ children }: { children: React.ReactNode }): JSX.Element {
   return (
     <div className="m-auto border-l-2 border-r-2 border-indigo-500 max-w-[360px] h-screen">
       <QueryProvider>{children}</QueryProvider>
