+++ apps/client/tailwind.config.ts
@@ -1,18 +1,5 @@
-import type { Config } from 'tailwindcss';
+/** @type {import('tailwindcss').Config} */
 import sharedConfig from '@wepl/tailwind-config';
-
-const config: Config = {
-  content: ['./app/**/*.{js,ts,jsx,tsx,mdx}', './src/**/*.{js,ts,jsx,tsx,mdx}'],
+module.exports = {
   presets: [sharedConfig],
-  theme: {
-    screens: {
-      '2xl': { max: '1535px' },
-      xl: { max: '1279px' },
-      lg: { max: '1023px' },
-      md: { max: '767px' },
-      sm: { max: '359px' },
-    },
-  },
 };
-
-export default config;
