+++ apps/client/tsconfig.json
@@ -5,7 +5,8 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./*"],
-      "@fsd/*": ["./src/*"]
+      "@fsd/*": ["./src/*"],
+      "@ui/*": ["../../packages/ui/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
