+++ packages/ui/components/ui/button.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
-
 import { cn } from '@ui/lib/utils';
 
 const buttonVariants = cva(
